+++ .eslintrc.json
@@ -5,6 +5,7 @@
     "semi": [0],
     "no-unused-vars": [0],
     "no-console": [0],
+    "no-plusplus": [0],
     "react/jsx-filename-extension": [0],
     "react/no-string-refs": [0],
     "jsx-a11y/tabindex-no-positive": [0]
